adding some conditionals, need to test

+++ new-org-account.ps1
@@ -216,19 +216,19 @@ function setup_guard_duty {
 
     $regions = @(
         @{
-            "region" = "us-east-2"
+            "region"     = "us-east-2"
             "detectorid" = "acb0ea346465917edef83687b7dfe06d"
         },
         @{
-            "region"    = "us-east-1"
+            "region"     = "us-east-1"
             "detectorid" = "0cb0f0c874d250b10e1dcee4cd168ffa"
         },
         @{
-            "region"    = "us-west-2"
+            "region"     = "us-west-2"
             "detectorid" = "deb0f1128a7c3e07e95961c01fa4c60e"
         },
         @{
-            "region"    = "us-west-1"
+            "region"     = "us-west-1"
             "detectorid" = "f4b100fe156d7207770c7bcc3268c3d5"
         }
     )
@@ -290,30 +290,40 @@ function delete_default_vpc {
     $regions = Get-AWSRegion
 
     $region | ForEach-Object -Process {
+        Write-Host "-------------------------"
+        Write-Host "Checking for Default VPCs"
+        Write-Host "-------------------------"
         $vpc = Get-EC2Vpc -Region $_.Region -Credential $Credentials -Filter @{Name = "isDefault"; Value = "true" }
-        $igw = Get-EC2InternetGateway -Region $_.Region -Credential $Credentials -Filter @{Name = "attachment.vpc-id"; Value = $vpc.VpcId }
-        Dismount-EC2InternetGateway -Region $_.Region -Credential $Credentials -VpcId $vpc.VpcId -InternetGatewayId $igw.InternetGatewayId
-        Remove-EC2InternetGateway -Region $_.Region -Credential $Credentials -InternetGatewayId $igw.InternetGatewayId
+        Write-Host "There are" ($vpc).count "Default VPCs in the Account"
 
-        $subnets = Get-EC2Subnet -Region $_.Region -Credential $Credentials -Filter @{Name = "vpc-id"; Value = $vpc.VpcId }
-        Remove-EC2Subnet -Region $_.Region -Credential $Credentials -SubnetId $subnets.SubnetId
+        if ($vpc.count -eq 0) {
+            Write-Host "There are no Default VPCs in" $_.Region
+        }
+        elseif ($vpc.count -gt 0) {
+            $igw = Get-EC2InternetGateway -Region $_.Region -Credential $Credentials -Filter @{Name = "attachment.vpc-id"; Value = $vpc.VpcId }
+            if ($igw) {
+                Write-Host "---------------------------------------------------------------"
+                Write-Host "Attempting to remove" $igw.InternetGatewayId "from VPC" $vpc.VpcId
+                Write-Host "---------------------------------------------------------------"
+                Dismount-EC2InternetGateway -Region $_.Region -Credential $Credentials -VpcId $vpc.VpcId -InternetGatewayId $igw.InternetGatewayId
+                Remove-EC2InternetGateway -Region $_.Region -Credential $Credentials -InternetGatewayId $igw.InternetGatewayId
+            }
+
+            $subnets = Get-EC2Subnet -Region $_.Region -Credential $Credentials -Filter @{Name = "vpc-id"; Value = $vpc.VpcId }
+            if ($subnets) {
+                Write-Host "---------------------------------------------"
+                Write-Host "Removing Subnets from Default VPC" $vpc.VpcId
+                Write-Host "---------------------------------------------"
+                Remove-EC2Subnet -Region $_.Region -Credential $Credentials -SubnetId $subnets.SubnetId
+            }
+        }
 
         Remove-EC2Vpc -VpcId $vpc.VpcId -Region $_.Region -Credential $Credentials -WhatIf
-        # $vpc | ForEach-Object -Process {
-        #     Remove-EC2Vpc -VpcId $_.VpcId -Region $_.Region -Credential $Credentials -WhatIf
-        # }
     }
 
 
 
 }
-
-
-
-
-
-
-
 
 
 function update_account_alias {
